% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_adoption.R
\docType{data}
\name{data_cat_adoption}
\alias{data_cat_adoption}
\alias{cat_adoption}
\title{Cat Adoption}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2257 rows and 18 columns.
}
\source{
\url{https://data.longbeach.gov/explore/dataset/animal-shelter-intakes-and-outcomes/information/}
}
\usage{
data_cat_adoption(...)

cat_adoption
}
\arguments{
\item{...}{Arguments passed to \code{\link[pins:pin_read]{pins::pin_read()}}.}
}
\value{
tibble
}
\description{
A subset of the cats at the animal shelter in Long Beach, California, USA.
}
\details{
A data frame with 2257 rows and 18 columns:
\describe{
\item{event_time}{The time the cat spent at the shelter. The event is
adoption or the cat being returned to the place they belong. The non-event is
the cat being transferred to another shelter or dying.}
\item{sex}{The sex of the cat and whether they are spayed or neutered.}
\item{intake_condition}{The intake condition of the cat.}
\item{intake_type}{The type of intake.}
\item{latitude}{Latitude of the intersection/cross street of intake or capture.}
\item{longitude}{Longitude of the intersection/cross street of intake or capture.}
\item{black}{Indicator for the color/pattern of the cat's fur.}
\item{brown}{Indicator for the color/pattern of the cat's fur.}
\item{brown_tabby}{Indicator for the color/pattern of the cat's fur.}
\item{calico}{Indicator for the color/pattern of the cat's fur.}
\item{cream}{Indicator for the color/pattern of the cat's fur.}
\item{gray}{Indicator for the color/pattern of the cat's fur.}
\item{gray_tabby}{Indicator for the color/pattern of the cat's fur.}
\item{orange}{Indicator for the color/pattern of the cat's fur.}
\item{orange_tabby}{Indicator for the color/pattern of the cat's fur.}
\item{tan}{Indicator for the color/pattern of the cat's fur.}
\item{tortie}{Indicator for the color/pattern of the cat's fur.}
\item{white}{Indicator for the color/pattern of the cat's fur.}
}
}
\section{tibble print}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{data_cat_adoption()
#> # A tibble: 2,257 x 18
#>    event_time sex    intake_condition intake_type latitude longitude black brown
#>        <Surv> <fct>  <fct>            <fct>          <dbl>     <dbl> <int> <int>
#>  1         17 neute~ fractious        owner_surr~     33.8     -118.     0     0
#>  2         98 neute~ normal           stray           33.8     -118.     0     0
#>  3         15 neute~ ill_moderatete   owner_surr~     33.8     -118.     0     0
#>  4         72 spayed fractious        owner_surr~     33.8     -118.     1     0
#>  5         22 neute~ normal           owner_surr~     33.8     -118.     0     0
#>  6         66 neute~ normal           owner_surr~     33.8     -118.     1     0
#>  7        200 spayed other            other           33.9     -118.     0     0
#>  8          9 spayed normal           owner_surr~     33.9     -118.     1     0
#>  9         45 neute~ ill_mild         stray           33.8     -118.     0     0
#> 10         38 male   ill_mild         stray           33.9     -118.     0     0
#>             1 sex    intake_condition intake_type latitude longitude black brown
#>             1 <fct>  <fct>            <fct>          <dbl>     <dbl> <int> <int>
#>  1          0 neute~ fractious        owner_surr~     33.8     -118.     0     0
#>  2          1 neute~ normal           stray           33.8     -118.     0     0
#>  3          0 neute~ ill_moderatete   owner_surr~     33.8     -118.     0     0
#>  4          1 spayed fractious        owner_surr~     33.8     -118.     1     0
#>  5          1 neute~ normal           owner_surr~     33.8     -118.     0     0
#>  6          0 neute~ normal           owner_surr~     33.8     -118.     1     0
#>  7          1 spayed other            other           33.9     -118.     0     0
#>  8          1 spayed normal           owner_surr~     33.9     -118.     1     0
#> # i 2,247 more rows
#> # i 10 more variables: brown_tabby <int>, calico <int>, cream <int>,
#> #   gray <int>, gray_tabby <int>, orange <int>, orange_tabby <int>, tan <int>,
#> #   tortie <int>, white <int>
}\if{html}{\out{</div>}}
}

\section{glimpse()}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{tibble::glimpse(data_cat_adoption())
#> Rows: 2,257
#> Columns: 18
#> $ event_time       <Surv> <Surv[26 x 2]>
#> $ sex              <fct> neutered, neutered, neutered, spayed, neutered, neut~
#> $ intake_condition <fct> fractious, normal, ill_moderatete, fractious, normal,~
#> $ intake_type      <fct> owner_surrender, stray, owner_surrender, owner_surren~
#> $ latitude         <dbl> 33.81065, 33.82919, 33.77225, 33.76832, 33.77914, 33.~
#> $ longitude        <dbl> -118.1036, -118.1730, -118.1710, -118.1520, -118.1718~
#> $ black            <int> 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0,~
#> $ brown            <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,~
#> $ brown_tabby      <int> 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,~
#> $ calico           <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,~
#> $ cream            <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,~
#> $ gray             <int> 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0,~
#> $ gray_tabby       <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0,~
#> $ orange           <int> 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1,~
#> $ orange_tabby     <int> 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0,~
#> $ tan              <int> 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,~
#> $ tortie           <int> 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,~
#> $ white            <int> 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,~
}\if{html}{\out{</div>}}
}

\examples{
\donttest{
data_cat_adoption()
}
}
\keyword{datasets}
