% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chicago_taxi.R
\name{data_chicago_taxi}
\alias{data_chicago_taxi}
\title{Chicago taxi data set}
\source{
\url{https://data.cityofchicago.org/Transportation/Taxi-Trips/wrvz-psew}
}
\usage{
data_chicago_taxi(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[pins:pin_read]{pins::pin_read()}}.}
}
\value{
tibble
}
\description{
A data set containing information on a subset of taxi trips in the city
of Chicago in 2022.
}
\details{
The source data are originally described on the linked City of Chicago
data portal. The data exported here are a pre-processed subset motivated by
the modeling problem of predicting whether a rider will tip or not.

\describe{
\item{tip}{Whether the rider left a tip. A factor with levels
"yes" and "no".}
\item{trip_id}{A unique identifier for the trip, as a factor.}
\item{trip_seconds}{The trip duration, in seconds.}
\item{trip_miles}{The trip distance, in miles.}
\item{fare}{The cost of the trip fare, in USD}
\item{tolls}{The cost of tolls for the trip, in USD.}
\item{extras}{The cost of extra charges for the trip, in USD.}
\item{trip_total}{The total cost of the trip, in USD. This is the
sum of the previous three columns plus tip.}
\item{payment_type}{Type of payment for the trip. A factor with levels
"Credit Card", "Dispute", "Mobile", "No Charge", "Prcard", and "Unknown".}
\item{company}{The taxi company, as a factor. Companies that occurred
few times were binned as "other".}
\item{local_trip}{Whether the trip started in the same community
area as it began. See the source data for community area values.}
\item{trip_start_dow}{The day of the week in which the trip began, as a
factor.}
\item{trip_start_month}{The month in which the trip began, as a factor.}
\item{trip_start_hour}{The hour of the day in which the trip began, as a
numeric.}
}
}
\section{tibble print}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{data_chicago_taxi()
#> # A tibble: 10,000 x 14
#>    tip   trip_id           trip_seconds trip_miles  fare tolls extras trip_total
#>    <fct> <fct>                    <dbl>      <dbl> <dbl> <dbl>  <dbl>      <dbl>
#>  1 yes   9fee331b5a4b19da~          333       1.24   6.5     0    0          8  
#>  2 no    b1dbc452aeb608f1~         2692       5.39  25.5     0    0         25.5
#>  3 yes   d5aaf396c67e55ff~         1076       3.01  11.2     0    0         14.1
#>  4 no    060c84fcdd4570be~         1599      18.4   45.2     0    0         45.2
#>  5 yes   716dbe81659166eb~          346       1.76   7.5     0    1         11  
#>  6 yes   617c96cf2a61cc6b~         1790      13.6   34.2     0    4.5       49.1
#>  7 yes   1e2d7bc1e6d30991~          885       3.71  11.2     0    0         12.6
#>  8 yes   7145ff35649e27b8~          720       4.8   14.8     0    1.5       19.6
#>  9 yes   ed985279be69b438~         2190      18.0   44.8     0    4         59.1
#> 10 no    730f737fa05eb908~         2160      17.5   44       0    4         48.5
#> # i 9,990 more rows
#> # i 6 more variables: payment_type <fct>, company <fct>, local_trip <fct>,
#> #   trip_start_dow <fct>, trip_start_month <fct>, trip_start_hour <int>
}\if{html}{\out{</div>}}
}

\section{glimpse()}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{tibble::glimpse(data_chicago_taxi())
#> Rows: 10,000
#> Columns: 14
#> $ tip              <fct> yes, no, yes, no, yes, yes, yes, yes, yes, no, yes, y~
#> $ trip_id          <fct> 9fee331b5a4b19daa19a149cfdfbeea91eb4dff9, b1dbc452aeb~
#> $ trip_seconds     <dbl> 333, 2692, 1076, 1599, 346, 1790, 885, 720, 2190, 216~
#> $ trip_miles       <dbl> 1.24, 5.39, 3.01, 18.38, 1.76, 13.65, 3.71, 4.80, 18.~
#> $ fare             <dbl> 6.50, 25.50, 11.23, 45.25, 7.50, 34.25, 11.25, 14.75,~
#> $ tolls            <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,~
#> $ extras           <dbl> 0.0, 0.0, 0.0, 0.0, 1.0, 4.5, 0.0, 1.5, 4.0, 4.0, 5.0~
#> $ trip_total       <dbl> 8.00, 25.50, 14.10, 45.25, 11.00, 49.06, 12.62, 19.60~
#> $ payment_type     <fct> Credit Card, Prcard, Mobile, Prcard, Credit Card, Cre~
#> $ company          <fct> Sun Taxi, Flash Cab, City Service, Sun Taxi, Sun Taxi~
#> $ local_trip       <fct> no, no, no, no, no, no, no, no, no, yes, no, NA, no, ~
#> $ trip_start_dow   <fct> Thu, Sat, Wed, Sat, Sun, Mon, Mon, Tue, Fri, Thu, Tue~
#> $ trip_start_month <fct> Feb, Mar, Feb, Apr, Jan, Feb, Mar, Mar, Jan, Apr, Apr~
#> $ trip_start_hour  <int> 13, 12, 17, 6, 15, 17, 21, 9, 19, 12, 20, 22, 10, 10,~
}\if{html}{\out{</div>}}
}

\examples{
\donttest{
data_chicago_taxi()
}
}
