% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevators.R
\name{data_elevators}
\alias{data_elevators}
\title{elevators data set}
\usage{
data_elevators()
}
\value{
tibble
}
\description{
elevators data set
}
\details{
The data set looks like:
}
\section{tibble print}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{data_elevators()
#> # A tibble: 35,042 x 25
#>    device_number bin     tax_b~1 tax_lot house~2 stree~3 zip_c~4 borough devic~5
#>    <chr>         <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <fct>   <chr>  
#>  1 1D10028       1024795 1021    26      1614    BROADW~ 10019   Manhat~ Dumbwa~
#>  2 1D10094       1041822 1392    25      53      E 77TH~ 10021   Manhat~ Dumbwa~
#>  3 1D10097       1038223 1323    1       201     E 49 ST 10017   Manhat~ Dumbwa~
#>  4 1D10146       1080443 1274    6       40      CENTRA~ <NA>    Manhat~ Dumbwa~
#>  5 1D10200       1085777 1074    24      651     TENTH ~ <NA>    Manhat~ Dumbwa~
#>  6 1D10301       1002075 181     16      179     FRANKL~ 10013   Manhat~ Dumbwa~
#>  7 1D10302       1010518 606     4       121     WEST 1~ 10011   Manhat~ Dumbwa~
#>  8 1D10303       1085955 1329    1       915     3 AVEN~ 10022   Manhat~ Dumbwa~
#>  9 1D10304       1044058 1430    5       220     E. 76 ~ 10021   Manhat~ Dumbwa~
#> 10 1D10305       1087468 1951    4       133     MORNIN~ <NA>    Manhat~ Dumbwa~
#> # ... with 35,032 more rows, 16 more variables: lastper_insp_date <date>,
#> #   approval_date <date>, manufacturer <chr>, travel_distance <chr>,
#> #   speed_fpm <dbl>, capacity_lbs <dbl>, car_buffer_type <chr>,
#> #   governor_type <chr>, machine_type <chr>, safety_type <chr>,
#> #   mode_operation <chr>, floor_from <chr>, floor_to <chr>, latitude <dbl>,
#> #   longitude <dbl>, elevators_per_building <int>, and abbreviated variable
#> #   names 1: tax_block, 2: house_number, 3: street_name, 4: zip_code, ...
}\if{html}{\out{</div>}}
}

\section{glimpse()}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{tibble::glimpse(data_elevators())
#> Rows: 35,042
#> Columns: 25
#> $ device_number          <chr> "1D10028", "1D10094", "1D10097", "1D10146", "1D~
#> $ bin                    <chr> "1024795", "1041822", "1038223", "1080443", "10~
#> $ tax_block              <chr> "1021", "1392", "1323", "1274", "1074", "181", ~
#> $ tax_lot                <chr> "26", "25", "1", "6", "24", "16", "4", "1", "5"~
#> $ house_number           <chr> "1614", "53", "201", "40", "651", "179", "121",~
#> $ street_name            <chr> "BROADWAY", "E 77TH ST", "E 49 ST", "CENTRAL PA~
#> $ zip_code               <chr> "10019", "10021", "10017", NA, NA, "10013", "10~
#> $ borough                <fct> Manhattan, Manhattan, Manhattan, Manhattan, Man~
#> $ device_type            <chr> "Dumbwaiter", "Dumbwaiter", "Dumbwaiter", "Dumb~
#> $ lastper_insp_date      <date> 2015-09-18, 2015-08-07, 2015-04-02, 2014-10-15~
#> $ approval_date          <date> 2006-03-07, 2006-05-15, 1998-09-21, 2010-08-02~
#> $ manufacturer           <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,~
#> $ travel_distance        <chr> "16'4\\"", NA, "23", "8'", "24 FT", "9'0", "12'0~
#> $ speed_fpm              <dbl> 50, 25, 50, 50, 50, 50, 50, 50, 50, 100, 100, 5~
#> $ capacity_lbs           <dbl> 500, 500, 500, 500, NA, 500, 300, 500, 500, 500~
#> $ car_buffer_type        <chr> "Spring", NA, NA, NA, NA, NA, "Spring", NA, NA,~
#> $ governor_type          <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,~
#> $ machine_type           <chr> NA, "OD", "BD", "BD", NA, "OD", "OD", "BD", "OG~
#> $ safety_type            <chr> "I", NA, "I", NA, NA, "I", "I", NA, "I", NA, NA~
#> $ mode_operation         <chr> "A", "P", "A", "A", NA, "A", "A", "A", "A", "P"~
#> $ floor_from             <chr> "B", "SB", "B", "B", "C", "BAS", "B", "C", "BMT~
#> $ floor_to               <chr> "1", "3", "2", "1", "G", "1", "1", "2", "4", "5~
#> $ latitude               <dbl> 40.76088, 40.77502, 40.75518, 40.76500, 40.7622~
#> $ longitude              <dbl> -73.98391, -73.96256, -73.97079, -73.97573, -73~
#> $ elevators_per_building <int> 11, 2, 1, 1, 2, 2, 1, 1, 1, 5, 5, 1, 2, 1, 1, 2~
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
data_elevators()
}
}
