% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeldatatoo_board.R
\name{internal_board}
\alias{internal_board}
\title{Internal pins board}
\usage{
internal_board()
}
\value{
a pins board
}
\description{
Pins board used internally to manage download, reading, and caching of data
sets.
}
\examples{
\donttest{
internal_board()
}

}
