% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeldatatoo-package.R
\docType{package}
\name{modeldatatoo-package}
\alias{modeldatatoo}
\alias{modeldatatoo-package}
\title{modeldatatoo: More Data Sets Useful for Modeling Examples}
\description{
More data sets used for demonstrating or testing model-related packages are contained in this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/modeldatatoo}
  \item \url{https://tidymodels.github.io/modeldatatoo/}
  \item Report bugs at \url{https://github.com/tidymodels/modeldatatoo/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emil.hvitfeldt@posit.co} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
