% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_adoption.R
\name{data_cat_adoption}
\alias{data_cat_adoption}
\title{Cat Adoption}
\source{
\url{https://data.longbeach.gov/explore/dataset/animal-shelter-intakes-and-outcomes/information/}
}
\usage{
data_cat_adoption(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[pins:pin_read]{pins::pin_read()}}.}
}
\value{
tibble
}
\description{
A subset of the cats at the animal shelter in Long Beach, California, USA.
}
\details{
A data frame with 2257 rows and 18 columns:
\describe{
\item{event_time}{The time the cat spent at the shelter. The event is
adoption or the cat being returned to the place they belong. The non-event is
the cat being transferred to another shelter or dying.}
\item{sex}{The sex of the cat and whether they are spayed or neutered.}
\item{intake_condition}{The intake condition of the cat.}
\item{intake_type}{The type of intake.}
\item{latitude}{Latitude of the intersection/cross street of intake or capture.}
\item{longitude}{Longitude of the intersection/cross street of intake or capture.}
\item{black}{Indicator for the color/pattern of the cat's fur.}
\item{brown}{Indicator for the color/pattern of the cat's fur.}
\item{brown_tabby}{Indicator for the color/pattern of the cat's fur.}
\item{calico}{Indicator for the color/pattern of the cat's fur.}
\item{cream}{Indicator for the color/pattern of the cat's fur.}
\item{gray}{Indicator for the color/pattern of the cat's fur.}
\item{gray_tabby}{Indicator for the color/pattern of the cat's fur.}
\item{orange}{Indicator for the color/pattern of the cat's fur.}
\item{orange_tabby}{Indicator for the color/pattern of the cat's fur.}
\item{tan}{Indicator for the color/pattern of the cat's fur.}
\item{tortie}{Indicator for the color/pattern of the cat's fur.}
\item{white}{Indicator for the color/pattern of the cat's fur.}
}
}
\section{tibble print}{
\if{html}{\out{<div class="sourceCode {r}">}}\preformatted{data_cat_adoption()
}\if{html}{\out{</div>}}
}

\section{glimpse()}{
\if{html}{\out{<div class="sourceCode {r}">}}\preformatted{tibble::glimpse(data_cat_adoption())
}\if{html}{\out{</div>}}
}

\examples{
\donttest{
data_cat_adoption()
}
}
