% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_adoption.R
\docType{data}
\name{data_cat_adoption}
\alias{data_cat_adoption}
\alias{cat_adoption}
\title{Cat Adoption}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2257 rows and 19 columns.
}
\source{
\url{https://data.longbeach.gov/explore/dataset/animal-shelter-intakes-and-outcomes/information/}
}
\usage{
data_cat_adoption(...)

cat_adoption
}
\arguments{
\item{...}{Arguments passed to \code{\link[pins:pin_read]{pins::pin_read()}}.}
}
\value{
tibble
}
\description{
A subset of the cats at the animal shelter in Long Beach, California, USA.
}
\details{
A data frame with 2257 rows and 19 columns:
\describe{
\item{event_time}{The time the cat spent at the shelter. The event is
adoption or the cat being returned to the place they belong. The non-event is
the cat being transferred to another shelter or dying.}
\item{sex}{The sex of the cat.}
\item{neutered}{Whether the cat is neutered.}
\item{intake_condition}{The intake condition of the cat.}
\item{intake_type}{The type of intake.}
\item{latitude}{Latitude of the intersection/cross street of intake or capture.}
\item{longitude}{Longitude of the intersection/cross street of intake or capture.}
\item{black}{Indicator for the color/pattern of the cat's fur.}
\item{brown}{Indicator for the color/pattern of the cat's fur.}
\item{brown_tabby}{Indicator for the color/pattern of the cat's fur.}
\item{calico}{Indicator for the color/pattern of the cat's fur.}
\item{cream}{Indicator for the color/pattern of the cat's fur.}
\item{gray}{Indicator for the color/pattern of the cat's fur.}
\item{gray_tabby}{Indicator for the color/pattern of the cat's fur.}
\item{orange}{Indicator for the color/pattern of the cat's fur.}
\item{orange_tabby}{Indicator for the color/pattern of the cat's fur.}
\item{tan}{Indicator for the color/pattern of the cat's fur.}
\item{tortie}{Indicator for the color/pattern of the cat's fur.}
\item{white}{Indicator for the color/pattern of the cat's fur.}
}
}
\section{tibble print}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{data_cat_adoption()
#> # A tibble: 2,257 x 19
#>    event_time sex    neutered intake_condition intake_type    latitude longitude
#>        <Surv> <fct>  <fct>    <fct>            <fct>             <dbl>     <dbl>
#>  1        17  male   yes      fractious        owner_surrend~     33.8     -118.
#>  2        98  male   yes      normal           stray              33.8     -118.
#>  3        15+ male   yes      ill_moderatete   owner_surrend~     33.8     -118.
#>  4        72  female yes      fractious        owner_surrend~     33.8     -118.
#>  5        22+ male   yes      normal           owner_surrend~     33.8     -118.
#>  6        66  male   yes      normal           owner_surrend~     33.8     -118.
#>  7       200  female yes      other            other              33.9     -118.
#>  8         9+ female yes      normal           owner_surrend~     33.9     -118.
#>  9        45  male   yes      ill_mild         stray              33.8     -118.
#> 10        38  male   no       ill_mild         stray              33.9     -118.
#> # i 2,247 more rows
#> # i 12 more variables: black <int>, brown <int>, brown_tabby <int>,
#> #   calico <int>, cream <int>, gray <int>, gray_tabby <int>, orange <int>,
#> #   orange_tabby <int>, tan <int>, tortie <int>, white <int>
}\if{html}{\out{</div>}}
}

\section{glimpse()}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{tibble::glimpse(data_cat_adoption())
#> Rows: 2,257
#> Columns: 19
#> $ event_time       <Surv> <Surv[26 x 2]>
#> $ sex              <fct> male, male, male, female, male, male, female, female~
#> $ neutered         <fct> yes, yes, yes, yes, yes, yes, yes, yes, yes, no, yes,~
#> $ intake_condition <fct> fractious, normal, ill_moderatete, fractious, normal,~
#> $ intake_type      <fct> owner_surrender, stray, owner_surrender, owner_surren~
#> $ latitude         <dbl> 33.81065, 33.82919, 33.77225, 33.76832, 33.77914, 33.~
#> $ longitude        <dbl> -118.1036, -118.1730, -118.1710, -118.1520, -118.1718~
#> $ black            <int> 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0,~
#> $ brown            <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,~
#> $ brown_tabby      <int> 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,~
#> $ calico           <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,~
#> $ cream            <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,~
#> $ gray             <int> 0, 0, 1, 0, 0, 1, 1, 0, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0,~
#> $ gray_tabby       <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0,~
#> $ orange           <int> 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1,~
#> $ orange_tabby     <int> 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0,~
#> $ tan              <int> 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,~
#> $ tortie           <int> 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,~
#> $ white            <int> 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,~
}\if{html}{\out{</div>}}
}

\examples{
\donttest{
data_cat_adoption()
}
}
\keyword{datasets}
